# Improves caching lookup performance