@nesquena 2025-07-15
> Would it be possible to cut a release with this change? 🙏

Yes I'll go ahead and push a release in the next week and respond back here